# docs/swagger.yaml
definitions:
...
  dto.AdminLoginInput:
    properties:
      password:
        description: 密码
        example: "123456"
        type: string
      username:
        description: 管理员用户名
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  dto.AdminLoginOutput:
    properties:
      token:
        description: token
        example: token
        type: string
    type: object
  dto.AssetAddInput:
    properties:
      area:
        type: string
      gps:
        type: string
      ip:
        type: string
      isp:
        type: string
      os:
        type: string
      task_id:
        type: integer
      vendor:
        type: string
    required:
    - ip
    - task_id
    type: object
  dto.AssetDetailOutput:
    properties:
      asset_info:
        $ref: '#/definitions/dto.AssetInfoCard'
        type: object
      box1:
        $ref: '#/definitions/dto.ChartBoxCard'
        type: object
      box2:
        $ref: '#/definitions/dto.ChartBoxCard'
        type: object
      trap_info:
        $ref: '#/definitions/dto.TrapInfo'
        type: object
      vul_list:
        $ref: '#/definitions/dto.VulTableBoxCard'
        type: object
      web_list:
        $ref: '#/definitions/dto.WebTableBoxCard'
        type: object
    type: object
  dto.AssetInfoCard:
    properties:
      area:
        type: string
      create_at:
        type: string
      gps:
        type: string
      id:
        type: integer
      image:
        description: 卡片头像
        type: string
      ip:
        type: string
      isp:
        type: string
      os:
        type: string
      target:
        type: string
      task_id:
        type: integer
      title:
        description: 卡片标题
        type: string
      type:
        description: 卡片绘图类型 pie line tatle
        type: string
      vendor:
        type: string
    type: object
  dto.AssetInfoOutput:
    properties:
      area:
        type: string
      create_at:
        type: string
      gps:
        type: string
      id:
        type: integer
      ip:
        type: string
      isp:
        type: string
      os:
        type: string
      port_array:
        items:
          type: string
        type: array
      task_id:
        type: integer
      task_name:
        type: string
      task_status:
        type: string
      vendor:
        type: string
      vul_count:
        type: integer
    type: object
  dto.AssetUpdateInput:
    properties:
      area:
        type: string
      gps:
        type: string
      id:
        type: integer
      ip:
        type: string
      isp:
        type: string
      os:
        type: string
      task_id:
        type: integer
      vendor:
        type: string
    required:
    - id
    type: object
...
  dto.ChartBoxCard:
    properties:
      image:
        description: 卡片头像
        type: string
      series:
        description: 数据源
        items:
          $ref: '#/definitions/dto.ChartSeries'
        type: array
      title:
        description: 卡片标题
        type: string
      type:
        description: 卡片绘图类型 pie line tatle
        type: string
    type: object
  dto.ChartSeries:
    properties:
      name:
        type: string
      value:
        type: integer
    type: object
...
  dto.TaskInfoOutput:
    properties:
      asset_num:
        type: integer
      auth_scan:
        type: integer
      create_at:
        type: string
      id:
        type: integer
      is_delete:
        type: integer
      name:
        type: string
      percent:
        type: integer
      poc_scan:
        type: integer
      progress:
        type: string
      recursion:
        type: integer
      rule_id:
        type: integer
      status:
        type: string
      target_list:
        type: string
      trap_scan:
        type: integer
      update_at:
        type: string
      web_scan:
        type: integer
    type: object
  dto.TaskListOutput:
    properties:
      list:
        items:
          $ref: '#/definitions/dto.TaskInfoOutput'
        type: array
      total:
        type: integer
    type: object
...
  dto.TokensInput:
    properties:
      grant_type:
        description: 授权类型
        example: client_credentials
        type: string
      scope:
        description: 权限范围
        example: read_write
        type: string
    required:
    - grant_type
    - scope
    type: object
  dto.TokensOutput:
    properties:
      access_token:
        description: access_token
        type: string
      expires_in:
        description: 超时时间
        type: integer
      scope:
        description: 权限范围
        type: string
      token_type:
        description: 类型
        type: string
    type: object
  dto.TrapInfo:
    properties:
      name:
        type: string
      verify:
        type: string
    type: object
...
  dto.VulInfoOutput:
    properties:
      app_name:
        type: string
      asset:
        type: string
      asset_id:
        type: integer
      create_at:
        type: string
      exploit_payload:
        type: string
      exploit_result:
        type: string
      exploit_url:
        type: string
      id:
        type: integer
      is_delete:
        type: integer
      line:
        type: integer
      plugin_id:
        type: integer
      port_id:
        type: integer
      trojan_payload:
        type: string
      trojan_result:
        type: string
      trojan_url:
        type: string
      verify_payload:
        type: string
      verify_result:
        type: string
      verify_url:
        type: string
      vul_name:
        type: string
      vul_type:
        type: string
      webshell_payload:
        type: string
      webshell_result:
        type: string
      webshell_url:
        type: string
    type: object
...
  dto.VulTableBoxCard:
    properties:
      image:
        description: 卡片头像
        type: string
      series:
        description: 数据源
        items:
          $ref: '#/definitions/dto.VulInfoOutput'
        type: array
      title:
        description: 卡片标题
        type: string
      type:
        description: 卡片绘图类型 pie line tatle
        type: string
    type: object
  dto.WebInfoOutput:
    properties:
      content_type:
        type: string
      id:
        type: integer
      login_list:
        type: string
      port_id:
        type: integer
      resource_list:
        type: string
      route_list:
        type: string
      server:
        type: string
      start_url:
        type: string
      sub_domain:
        type: string
      title:
        type: string
      upload_list:
        type: string
    type: object
  dto.WebTableBoxCard:
    properties:
      image:
        description: 卡片头像
        type: string
      series:
        description: 数据源
        items:
          $ref: '#/definitions/dto.WebInfoOutput'
        type: array
      title:
        description: 卡片标题
        type: string
      type:
        description: 卡片绘图类型 pie line tatle
        type: string
    type: object
  middleware.Response:
    properties:
      data:
        type: object
      errmsg:
        type: string
      errno:
        type: integer
      stack:
        type: object
      trace_id:
        type: object
    type: object
info:
  contact: {}
  license: {}
paths:
...
  /admin_login/login:
    post:
      consumes:
      - application/json
      description: 管理员登陆
      operationId: /admin_login/login
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AdminLoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AdminLoginOutput'
              type: object
      summary: 管理员登陆
      tags:
      - 管理员接口
...
  /asset/add:
    post:
      consumes:
      - application/json
      description: 资产添加
      operationId: /asset/add
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AssetAddInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 资产添加
      tags:
      - 资产管理
  /asset/delete:
    delete:
      consumes:
      - application/json
      description: 资产删除
      operationId: /asset/delete
      parameters:
      - description: 资产ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 资产删除
      tags:
      - 资产管理
  /asset/detail:
    get:
      consumes:
      - application/json
      description: 资产详情
      operationId: /asset/detail
      parameters:
      - description: 资产ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AssetDetailOutput'
              type: object
      summary: 资产详情
      tags:
      - 资产管理
  /asset/info:
    get:
      consumes:
      - application/json
      description: 资产信息
      operationId: /asset/info
      parameters:
      - description: 资产ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AssetInfoOutput'
              type: object
      summary: 资产信息
      tags:
      - 资产管理
  /asset/list:
    get:
      consumes:
      - application/json
      description: 资产列表
      operationId: /asset/list
      parameters:
      - description: 模糊查询
        in: query
        name: info
        type: string
      - description: 每页条数
        in: query
        name: limit
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.TaskListOutput'
              type: object
      summary: 资产列表
      tags:
      - 资产管理
  /asset/update:
    put:
      consumes:
      - application/json
      description: 资产更新
      operationId: /asset/update
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AssetUpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 资产更新
      tags:
      - 资产管理
...
  /oauth/toekns:
    post:
      consumes:
      - application/json
      description: 获取TOKEN
      operationId: /oauth/toekns
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.TokensInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.TokensOutput'
              type: object
      summary: 获取TOKEN
      tags:
      - OAUTH
...
swagger: "2.0"